+++ AllMyLights/Models/Transformations/ExpressionTransformationOptions.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace AllMyLights.Models.Transformations
+{
+    public class ExpressionTransformationOptions: TransformationOptions
+    {
+        [Required]
+        public string Expression { get; set; }
+    }
+}